requirements.txt
Django
-e git+https://github.com/Marx86/django-voting.git#egg=django-voting
django-taggit
django-mptt
pytz
Pillow
django-hvad
python-social-auth
django-coffeescript
BeautifulSoup
-e git+https://github.com/Marx86/supercaptcha.git#egg=supercaptcha
django-imagekit
psycopg2
django-registration-redux
